WEEK_8 - Django_Framwork(1)/Day-2/Challenge_day-2/person/models.py
from django.db import models
from phonenumber_field.modelfields import PhoneNumberField

# Create your models here.

class Person(models.Model):
    name = models.CharField( max_length=50)
    email = models.CharField(max_length=50)
    phone = PhoneNumberField()
    adresse = models.CharField(max_length=50)
       

    def __str__(self):
        return self.name


    