WEEK_8 - Django_Framwork(1)/Day-2/Challenge_day-2/requirements.txt
asgiref==3.6.0
Django==4.1.5
django-phonenumber-field==7.0.1
phonenumbers==8.13.4
phonenumberslite==8.13.4
sqlparse==0.4.3
